requirements.txt
amqp==5.3.1
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
billiard==4.2.1
celery==5.5.1
certifi==2025.1.31
cfgv==3.4.0
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
distlib==0.3.9
Django==5.2
django-cors-headers==4.7.0
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
executing==2.2.0
filelock==3.18.0
h11==0.16.0
httptools==0.6.4
identify==2.6.9
idna==3.10
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kombu==5.5.3
loguru==0.7.3
matplotlib-inline==0.1.7
msgpack==1.1.0
nest-asyncio==1.6.0
nodeenv==1.9.1
numpy==2.2.5
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
pre_commit==4.2.0
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg==3.2.6
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.1
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-ldap==3.4.4
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
redis==5.2.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
vine==5.1.0
virtualenv==20.30.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
